+++ server/messages/pkg/model/message.go
@@ -6,13 +6,13 @@ import (
 )
 
 type Message struct {
-	ID                  int32               `json:"id,omitempty"`
-	ThreadID            int32               `json:"thread_id,omitempty"`
+	ID                  int64               `json:"id,omitempty"`
+	ThreadID            int64               `json:"thread_id,omitempty"`
 	CreateUTCNano       int64               `json:"create_utc_nano,omitempty"`
 	UpdateUTCNano       int64               `json:"update_utc_nano,omitempty"`
-	UserID              int32               `json:"user_id,omitempty"`
-	FileID              int32               `json:"file_id,omitempty"`
-	FileIDs             []int32             `json:"file_ids,omitempty"`
+	UserID              int64               `json:"user_id,omitempty"`
+	FileID              int64               `json:"file_id,omitempty"`
+	FileIDs             []int64             `json:"file_ids,omitempty"`
 	Name                string              `json:"name,omitempty"`
 	File                *filesmodel.File    `json:"file,omitempty"`
 	Text                string              `json:"text"`
@@ -36,30 +36,30 @@ type NewMessageServerResponse struct {
 
 type UpdateMessageServerResponse struct {
 	model.ServerResponse
-	ID                  int32         `json:"id"`
+	ID                  int64         `json:"id"`
 	UpdateUTCNano       int64         `json:"update_utc_nano"`
 	Private             bool          `json:"private"`
 }
 
 type DeleteMessageServerResponse struct {
 	model.ServerResponse
-	ID                  int32         `json:"id"`
+	ID                  int64         `json:"id"`
 }
 
 type PublishMessagesServerResponse struct {
 	model.ServerResponse
-	IDs                  []int32       `json:"ids"`
+	IDs                  []int64       `json:"ids"`
 	UpdateUTCNano        int64         `json:"update_utc_nano"`
 }
 
 type PrivateMessagesServerResponse struct {
 	model.ServerResponse
-	IDs                  []int32       `json:"ids"`
+	IDs                  []int64       `json:"ids"`
 	UpdateUTCNano        int64         `json:"update_utc_nano"`
 }
 
 type DeleteMessageStatus struct {
-	ID                int32           `json:"id"`
+	ID                int64           `json:"id"`
 	OK                bool            `json:"ok"`
 	Explain           string          `json:"explain"`
 }
@@ -71,7 +71,7 @@ type DeleteMessagesServerResponse struct {
 
 type MessagesListServerResponse struct {
 	model.ServerResponse
-	ThreadID            int32         `json:"thread_id"`
+	ThreadID            int64         `json:"thread_id"`
 	Messages            []*Message    `json:"messages"`
 	IsLastPage          bool          `json:"is_last_page"`
 }
